requirements.txt
accelerate==0.17.1
bitsandbytes==0.38.0.post1
controlnet_aux==0.0.3
datasets==2.15.0
decord==0.6.0
diffusers==0.25.1
dominate==2.9.1
einops==0.8.0
huggingface_hub==0.20.2
imageio==2.9.0
insightface==0.2.1
jsonlines==4.0.0
matplotlib==3.7.1
moviepy==1.0.3
numpy==1.23.1
onnx==1.16.1
onnxruntime-gpu==1.18.1
opencv_contrib_python==4.3.0.36
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
packaging==24.0
Pillow==10.3.0
positional_encodings==6.0.1
pytorch_lightning==1.5.0
safetensors==0.4.3
scipy==1.9.1
seaborn==0.13.2
tensorflow
torch
torchvision
tqdm==4.65.0
transformers==4.27.1
wandb==0.14.0
xformers==0.0.23.post1
